week08-react/minseon/requirements.txt
google-generativeai>=0.8.0
requests>=2.31.0
python-dotenv>=1.0.0
anthropic>=0.40.0
streamlit>=1.40.0
